- yet more tweaks to the gene overlap code, covering more and more obscure cases   - far parts not part of determining if mixed strand   - keep filtering gene_xref in mind, and find best that matches xref, if any   - etc. - don't trim initial methionine if the protein chain consists of just ONE amino   acid, which is a met - a little code cleanup (nothing major)   - code paths that are impossible to execute   - calls to deprecated functions   - some compiler warnings - use /codon_recognized tag (NOT note) in gbench mode JIRA GP-1267
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@49685 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/format/utils.cpp
@@ -532,7 +532,7 @@ bool TrimSpacesAndJunkFromEnds(string& str, bool allow_ellipsis)
 
     // check for ';' that's part of an HTML escape char like "&bgr;" and
     // skip over it (i.e., don't remove it) if so
-    if( start_of_junk_pos < str.length() && str[start_of_junk_pos] == ';' ) {
+    if( start_of_junk_pos < (int)str.length() && str[start_of_junk_pos] == ';' ) {
         // we assume no HTML escape char will be longer than this
         static const int kMaxCharsToLookAt = 20;
 
@@ -587,7 +587,7 @@ bool TrimSpacesAndJunkFromEnds(string& str, bool allow_ellipsis)
                 str += suffix;
                 changed = true;
             }
-        } else if ( start_of_junk_pos < str.length() ) {
+        } else if ( start_of_junk_pos < (int)str.length() ) {
             str.erase( start_of_junk_pos );
             changed = true;
         }
